v0.2.0: fix Tibber API schema, drop Docker build step, add logging
Tibber API changes:
- Replace homes[].vehicles with me.myVehicles.vehicles (schema changed)
- Drop appNickname from Home query (field removed from app API)
- Add shortName alongside title for vehicle display names

Volvo API fixes:
- Decode electricRange.value as float64 before truncating to int
  (API returns 53.0, not 53)

Monitor improvements:
- Log "poll starting" at the top of each cycle
- Log Volvo and Tibber token refreshes on success

Docker / deployment:
- docker-compose.yml: mount ./monitor binary instead of building from source;
  use gcr.io/distroless/static-debian12 as the runtime image — no Go toolchain
  needed on the Pi, upgrades are a binary swap + restart
- HTTP_ADDR now defaults to empty (server disabled); set to e.g. :8080 to
  enable /healthz and /state endpoints
- release.sh: build binaries with Go toolchain directly instead of
  nonexistent Dockerfile.build
- install.sh: remove --build flag from docker compose commands

Docs:
- README: rewrite Volvo portal setup to match current UI (generate client
  secret via Application Client Details, no manual redirect URI registration)
- README: document optional HTTP endpoints, VEHICLE_NOT_FOUND troubleshooting

+++ internal/volvo/client.go
@@ -48,8 +48,8 @@ type resourceFloat struct {
 }
 
 type resourceInt struct {
-	Status string `json:"status"`
-	Value  int    `json:"value"`
+	Status string  `json:"status"`
+	Value  float64 `json:"value"`
 }
 
 type resourceString struct {
@@ -107,11 +107,11 @@ func (c *Client) FetchChargeState(ctx context.Context) ChargeState {
 		state.BatteryChargeLevelPct = &v
 	}
 	if e.ElectricRange.Status == "OK" {
-		v := e.ElectricRange.Value
+		v := int(e.ElectricRange.Value)
 		state.ElectricRangeKm = &v
 	}
 	if e.EstimatedChargingTimeToTargetBatteryCharge.Status == "OK" {
-		v := e.EstimatedChargingTimeToTargetBatteryCharge.Value
+		v := int(e.EstimatedChargingTimeToTargetBatteryCharge.Value)
 		state.EstimatedChargingTimeMin = &v
 	}
 	if e.ChargingStatus.Status == "OK" {
